+++ src/HexagoScreenSaverConfig.cpp
@@ -12,15 +12,15 @@ namespace hexago {
 
     // main constructor
     HexagoScreenSaverConfig::HexagoScreenSaverConfig(
-        ParameterRange<float> start_size_range,
-        ParameterRange<float> decay_speed_range,
+        ParameterRange<double> start_size_range,
+        ParameterRange<double> decay_speed_range,
         colour_model_t colour_model,
         ParameterRange<colour_channel_t> d_colour_channel_range,
         ParameterRange<colour_channel_t> e_colour_channel_range,
         ParameterRange<colour_channel_t> f_colour_channel_range,
         ParameterRange<colour_channel_t> alpha_colour_channel_range,
         unsigned int framerate,
-        float minimum_screen_cover,
+        double minimum_screen_cover,
         hexagon_spawn_mode_t spawn_mode,
         background_colour_t background_colour
     )
@@ -45,7 +45,7 @@ namespace hexago {
     HexagoScreenSaverConfig::HexagoScreenSaverConfig(
         HexagonFactoryConfig sub_config,
         unsigned int framerate,
-        float minimum_screen_cover,
+        double minimum_screen_cover,
         hexagon_spawn_mode_t spawn_mode,
         background_colour_t background_colour
     )
